assessment_judge/requirements.txt
anthropic==0.39.0
requests==2.32.3
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
